separation of server configurations from its execution
